+++ assets/javascripts/kitten/components/layout/dashboard-layout/index.js
@@ -332,6 +332,32 @@ const Alerts = ({ className, ...props }) => (
   />
 )
 
+const Toaster = ({ className, isOpen, children, ...props }) => {
+
+  return (
+    <>
+      <div
+        className={classNames(
+          'k-DashboardLayout__toaster__wrapper',
+          'k-DashboardLayout__fullWidth',
+          className,
+          {
+            'k-DashboardLayout__toaster--isOpen': isOpen,
+          }
+        )}
+        aria-live="polite"
+        {...props}
+      >
+        <div className="k-DashboardLayout__toaster">
+          {children}
+        </div>
+      </div>
+      <div className="k-DashboardLayout__toaster__spacer" />
+    </>
+  )
+}
+
+
 DashboardLayout.propTypes = {
   backLinkProps: PropTypes.object,
   buttonProps: PropTypes.shape({
@@ -357,3 +383,4 @@ DashboardLayout.SideContent = SideContent
 DashboardLayout.SideFooter = SideFooter
 DashboardLayout.Flow = Flow
 DashboardLayout.Alerts = Alerts
+DashboardLayout.Toaster = Toaster